doc: entity-tag (#116)
An entity-tag (ETAG) is used to identify changes of the resource that happen over time. ETAG is generated by the resource server, in our case device, using various methods such as a version, checksum, hash or a timestamp. When the client retrieves an entity-tag, it should treat it as an opaque value without making any assumptions about its content or structure. In CoAP, the ETAG is part of the [CoAP options](https://datatracker.ietf.org/doc/html/rfc7252#section-5.4) and is used to detect changes in resources. It is used to optimize the client processing time of the response from the device in case no change was detected - the ETAG wasn't changed from the last request.